+++ src/libsac2c/global/private_heap.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-extern heap *PHPcreateHeap (size_t elem_size, int chunk_size);
+extern heap *PHPcreateHeap (size_t elem_size, size_t chunk_size);
 extern void *PHPmalloc (heap *private_heap);
 extern void *PHPfindElem (heap *private_heap, php_cmp_fun fun, void *elem);
 extern heap *PHPfreeHeap (heap *private_heap);
