+++ src/libsac2c/tree/tree_compound.h
@@ -77,12 +77,12 @@ extern node *TCshpseg2Array (shpseg *shape, int dim);
 
 extern types *TCappendTypes (types *chain, types *item);
 extern int TCcountTypes (types *type);
-extern types *TCgetTypesLine (types *type, int line);
+extern types *TCgetTypesLine (types *type, size_t line);
 extern types *TCgetTypes (types *type);
 extern int TCgetShapeDim (types *type);
 extern int TCgetDim (types *type);
 extern simpletype TCgetBasetype (types *type);
-extern int TCgetBasetypeSize (types *type);
+extern size_t TCgetBasetypeSize (types *type);
 extern int TCgetTypesLength (types *type);
 extern shpseg *TCtype2Shpseg (types *type, int *ret_dim);
 extern shape *TCtype2Shape (types *type);
@@ -695,7 +695,7 @@ extern node *TCmakeExprsNum (int num);
  *   Computes the length of the given N_exprs chain.
  *
  ******************************************************************************/
-
+/* FIXME may be better to make this unsigned int in the future */
 extern int TCcountExprs (node *exprs);
 
 extern node *TCgetNthExprs (int n, node *exprs);
